Added -Wconversion more flags to warnings.
Don't use -Wno-unused but -Wno-unused-parameter instead. Because
we don't want to be harassed with unused function parameters but
with unused variable we do. (gcc3 only)

Changed the behaviour for the warning flags: By default -W -Wall
-Wconversion are active. All warnings produced with this settings
SHOULD BE FIXED!!!

In certain cases it could be useful to turn on even more warnings by
using --enable-more-warnings. But be warned, with this setting gcc
becomes really rebellious. :-)

Pass --enable-more-warnings to autogen.sh or configure.
You may get a lot of false alarms depending on your compiler version.

-enable-more-warnings activates -Wshadow -Wunreachable-code (and more
to come)


git-svn-id: https://www.sat.ltu.se/svn/rt/arts/trunk@2391 aaf1aab0-4228-0410-ad68-8dceda47f409