* arts-1-1-594
* src/matpackI.h (Range): Added selectors for start, extent and range
to avoid unnecessary need for friend declarations.
Removed friend declaration of Sparse output operator, transpose and
mult.

* src/matpackII.{cc,h} (operator<<, mult, transpose): Use selector
functions instead of direct member access. This has a little impact
on speed when compiling without optimizations and not noticeable with
-O2 and higher.

The operator() is for read-only access. To write elements the rw()
modifier has to be used. Otherwise, e.g cout << sparse(1,1); will
create the element instead of just outputting it.

* src/m_sensor.cc (sensor_responseInit): Use rw() instead of
operater().

* src/xml_io_basic_types.cc (xml_write_to_stream): Use selector
functions instead of direct access for the Sparse range member
attributes.


git-svn-id: https://www.sat.ltu.se/svn/rt/arts/trunk@3256 aaf1aab0-4228-0410-ad68-8dceda47f409