* arts-1-1-627
* src/methods.cc, src/m_general.cc: Added dummy Print method for
group Timer.
Added methods timerStart and timerStop.

* src/m_general.h: Added dummy Print method for group Timer.
Fixed filename in doxygen header.

* src/workspace.cc: Added WSV timer.

* src/groups.cc: Added group Timer.

* src/make_auto_wsv_groups_h.cc: Include m_general.h in
auto_wsv_groups.h.

* src/make_auto_wsv_h.cc: Include m_general.h in auto_wsv.h.

* src/Makefile.am: make_auto_wsv_h and make_auto_wsv_group_h must
depend on m_general.h.

* src/xml_io_basic_types.{h,cc}, src/xml_io_instantiation.h: Added
dummy read/write functions for group Timer.

* doc/doxygen/html/Makefile.am: Added graph_legend.dot and
formula.repository to MAINTAINERCLEANFILES.

* doc/doxygen/Makefile.am: Use CLEANFILES instead of target clean.
Otherwise 'doxyfied' is not removed with make maintainer-clean.
Readded latex to subdirs.

* doc/doxygen/latex/.cvsignore: Added refman.out.

* doc/doxygen/latex/Makefile.am: Define empty 'all' target as we don't
won't to build ps/dvi by default but the user should be able to do so
if she likes.
Undef EXTRA_DIST.

* doc/doxygen/Doxyfile.in: Disabled creation of man pages and rtf. If
we want to support this, we should also create man and rtf directories
with appropriate Makefiles in the repository.

* configure.in: Readded doc/doxygen/latex/Makefile to AC_OUTPUT. We
don't build latex doxygen documentation by default, but it should be
possible to do it manually. For this case we need the Makefile.

* src/gas_abs_lookup.cc: Put empty doxygen header in front of output
operator to prevent doxygen confusion.

* doc/doc++/{.,html,tex}/.cvsignore: Removed to make directories empty
and pruned on checkout/update.

* ChangeLog: Wrapped paragraphs with lines > 80 chars. Cory, your
emacs sucks ;-)


git-svn-id: https://www.sat.ltu.se/svn/rt/arts/trunk@3462 aaf1aab0-4228-0410-ad68-8dceda47f409