* arts-1-1-941
* src/matpackI.{h,cc}: Added two new VectorView constructors.

(VectorView::VectorView(const Vector&)): Fail immediately if someone
tries to create a VectorView from a const Vector.

(VectorView::VectorView(Vector&)): Added to allow the creation of a
VectorView from a non-constant Vector. Without this constructor, the
VectorView::VectorView(const Vector&) will always be called which is
not what we want.

* src/cloudbox.{cc,h} (chk_single_scattering_data):

Use ConstVectorView instead of 'const VectorView'.


git-svn-id: https://www.sat.ltu.se/svn/rt/arts/trunk@4381 aaf1aab0-4228-0410-ad68-8dceda47f409