* arts-1-1-1129
* tests/Makefile.am:

Added a check-parallel target so the testing framework can benefit
from multiple processor systems.
E.g. 'make -j4 check-parallel' to run on 4 processors.
To keep the output readable (as it might mix on parallel systems), it
is written to a log file for each test.

TODO: Refurbish/shorten the target rules.

* test/.cvsignore: Ignore log files.


git-svn-id: https://www.sat.ltu.se/svn/rt/arts/trunk@4805 aaf1aab0-4228-0410-ad68-8dceda47f409