* arts-1-1-1188
* configure.in: Check if POSIX functions are available.

* src/m_general.{cc,h}:

Timers are only available if the build environment has POSIX support.

* doc/uguide/Figs/Makefile.am:

Loop over the list of figure directories and add the pdfs in each
directory separately to EXTRA_DIST. With the */*.pdf wildcard I used
before we end up with an extra directory called '*' in the
distribution archive which is not what we want.


git-svn-id: https://www.sat.ltu.se/svn/rt/arts/trunk@5489 aaf1aab0-4228-0410-ad68-8dceda47f409