* arts-1-3-17
* Continued work on the new controlfile features.

WSM input and output variables are now treated the same as generic
inputs and outputs.
In practice, this means that also for non-generic inputs and outputs
arbitrary workspace variables (of matching type) can be passed to the
method in the controlfile.

* src/parser.{cc,h}:

(parse_input): Add input variable array to the parameter list.

(parse_output): Add output variable array to the parameter list.

Indexes for the actual variables passed as input and output are now
added to the MRecord's moutput and minput arrays. Before, this was
only done for generic variables.

* src/make_auto_md_cc.cc:

Handle input and output WSVs in the same way as generic variables.

* src/agenda_class.h (MRecord): Updated docs for minput/moutput.


git-svn-id: https://www.sat.ltu.se/svn/rt/arts/trunk@5657 aaf1aab0-4228-0410-ad68-8dceda47f409