* arts-1-4-11
* src/arts_omp.cc (arts_omp_in_parallel):

omp_in_parallel behaves differently in gcc and icc. omp_in_parallel
in the intel compiler returns true even if the maximum number of
threads is set to 1, but false in gcc. On the other hand,
omp_get_max_threads in gcc always returns 1 instead of the correct
number of max threads when called inside a parallelized region. So,
we have to use to use omp_get_max_threads with the intel compiler and
omp_in_parallel with gcc to determine if we're running multiple threads
or not. Bummer.


git-svn-id: https://www.sat.ltu.se/svn/rt/arts/trunk@5696 aaf1aab0-4228-0410-ad68-8dceda47f409