* arts-1-9-22
* src/methods_aux.cc:

Part One of the online help parameter documentation beautification.

Pull the short description for each WSM's input/output variable from
the WSV documentation and display it in the WSM's online help.

Also change the way the variables' documentation is formatted in the
WSM online help.

Documentation for GIN/GOUT will follow in Part 2.


git-svn-id: https://www.sat.ltu.se/svn/rt/arts/trunk@5867 aaf1aab0-4228-0410-ad68-8dceda47f409