* arts-1-9-25
* src/agenda_class.cc (Agenda::execute):

Fix serious bug discovered by Patrick. Make sure variables used both
as input and output are initialized.

* src/parser.cc (ArtsParser::parse_generic_input):

Instead of crashing, throw an error and tell the user to use Copy if
a variable instead of a constant is passed to a Set method.

* src/methods.h, src/methods_aux.cc:

Added function to get a list of variables that are used both as input
and output.


git-svn-id: https://www.sat.ltu.se/svn/rt/arts/trunk@5871 aaf1aab0-4228-0410-ad68-8dceda47f409