* arts-1-13-25
* CMakeLists.txt, cmake/modules/CheckOpenMPThreadprivate.cmake:

Instead of disabling OpenMP altogether if threadprivate is not
supported, just don't use it. This makes it possible to use OpenMP on
Mac OS X. Only drawback is not so pretty agenda output which is better
than not having OpenMP at all.

We don't check for OpenMP support if we generate an Xcode project
file. Adding our own OpenMP compiler flags will break compilation, so
we leave all that to Xcode.

* config.h.cmake, src/messages.{cc,h}, src/main.cc:

Use new preprocessor variable THREADPRIVATE_SUPPORTED.

* doc/uguide/main.tex: Update christian's email.

* .: Clean up svn:ignore property.


git-svn-id: https://www.sat.ltu.se/svn/rt/arts/trunk@6299 aaf1aab0-4228-0410-ad68-8dceda47f409