* arts-1-14-25
* cmake/modules/ArtsBuildTexDoc.cmake:

Split building of each document into two stages. The first stage runs
latex, makeindex, bibtex, and latex. After that the aux files are
copied to the build directories of the other documents. The second
stage is another latex run for each document. Now it's possible to
parallel build the docs and get the cross references right.


git-svn-id: https://www.sat.ltu.se/svn/rt/arts/trunk@6352 aaf1aab0-4228-0410-ad68-8dceda47f409