* src/gas_abs_lookup.cc (GasAbsLookup::Extract):
Change the local GridPosPoly variables to ArrayOfGridPosPoly with just
one element. That avoids the creation of a temporary array on every
call of gridpos_poly. Depending on the control file this speeds up
absorption calculation by up to 10%.

* src/interpolation_poly.cc (gridpos_poly):

Pass the Numeric new_grid directly to other gridpos_poly function
instead of creating a temporary vector.


git-svn-id: https://www.sat.ltu.se/svn/rt/arts/trunk@6378 aaf1aab0-4228-0410-ad68-8dceda47f409