* arts-1-14-101
* src/docserver.{h,cc}:

Add baseurl support to allow integration into a webserver.
On the SAT server, we use the apache proxy module to redirect
all requests to http://www.sat.ltu.se/arts/docserver to the docserver.
In this case arts has to be started with
./arts -S -B /arts/docserver so it can preceed all links it generates
with the proper baseurl.

Output arts version in footer.

* src/main.cc, src/parameters.{h,cc}:

Add option --baseurl.


git-svn-id: https://www.sat.ltu.se/svn/rt/arts/trunk@6603 aaf1aab0-4228-0410-ad68-8dceda47f409