* arts-1-14-190
* src/m_doit.cc (DoitInit), src/methods.cc:

Remove doit_za_interp from output list. d_z_i was defined as
output-only but the initialization code relied upon its current value.

if (doit_za_interp != 1) doit_za_interp = 0;

d_z_i was effectively ending up with a random value of 0 or 1.

On Linux this was not visible as uninitialized variables seem to have
a tendency to have 0 as a value. But this is in no way guaranteed.
This led to a very hard-to-hunt-down bug that showed up in only 5-10%
of DOIT calculations and mostly only on Mac OS X.

The ugly bit was that even though doit_za_interpSet is properly
called in the case of TestDOITBatch, doit_za_interp would still end up
with a random value of 0 or 1 if DoitInit is called inside an
agenda. Because DoitInit was defined to generate doit_za_interp,
the scoping mechanism would not copy the original value of
doit_za_interp. There is no need to do that for variables that are
supposed to be generated by a WSM inside the agenda anyway.
Which made it even more difficult to track down was the fact that the
random change of the used interpolation from linear to polynomial only
showed differences in the results after around 55 Doit iterations.

Moral of the story:

If you define a workspace variable as output-only for a workspace
method, you MUST INITIALIZE IT inside the WSM. There is no way around
it.

* src/methods.cc (DoitCloudFieldPut):

Same problem here with scat_i_p. Defined as output-only, but accessed
with issize(scat_i_p,....) in the 1D case.
I've changed it to be input and output for now. But I'm not sure if
this is the right fix. Maybe instead of issize() there should be a
resize(scat_i_p...) like in the 3D case?

* src/math_funcs.cc (lgamma_func):

Make static coefficient array also const.

Reindented file.

* src/m_abs_o2_models.cc (Zeeman_o2_line_splitting):

Add warning to doc that this function is not threadsafe.


git-svn-id: https://www.sat.ltu.se/svn/rt/arts/trunk@6895 aaf1aab0-4228-0410-ad68-8dceda47f409