* arts-2-1-16
* src/m_abs.cc (abs_linesReadFromSplitArtscat):

If basename is a directory (path ends with '/'), don't append
'.SPECIES.xml' but just 'SPECIES.xml'.

* src/xml_io_array_types.cc (xml_read_from_stream):

Fix regression for reading ArrayOfLineRecord from an ARTS catalog whose
version number is just "3" instead of "ARTSCAT-3".

(xml_write_to_stream): When writing an ArrayOfLineRecord, determine
the required catalog version from the first LineRecord. If the array
contains a mixture of different versions of lines, throw a runtime
error.

* src/xml_io.cc (xml_read_from_file,
xml_read_arts_catalogue_from_file):

Output the _full_ filename to out2. Gives better feedback to the user
if the file is found somewhere in the path (ARTS_DATA_PATH) instead of
the current directory.


git-svn-id: https://www.sat.ltu.se/svn/rt/arts/trunk@7386 aaf1aab0-4228-0410-ad68-8dceda47f409