* arts-2-1-52
* src/m_atmosphere.cc (AtmFieldsCalc):

Add 1D support for VMR input profiles that do not cover the whole
range of the target grid. Boundary values in the partial input profile
must be 0.

* src/check_input.{cc,h} (chk_interpolation_grids_loose):

Add new check used by AtmFieldsCalc to determine the range of the
new grid that is covered by the old grid. Also checks that boundary
values are 0 in the input data.

* src/m_basic_types.cc, src/methods.cc (VectorCompare, MatrixCompare):

Make implementation and error output of both methods consistent. Add
an optional String argument for passing a custom error message.


git-svn-id: https://www.sat.ltu.se/svn/rt/arts/trunk@7456 aaf1aab0-4228-0410-ad68-8dceda47f409