* arts-2-1-59
* CMakeLists.txt, src/CMakeLists.txt, tests/CMakeLists.txt:

Started utilizing the ctest facility of cmake. Should make it much
easier to integrate tests in the future. Also, the output looks very
clean compared to now.

If you want to give it a try, run 'make newcheck' in the toplevel of
the build directory. Note that only one controlfile is run at the
moment. More to come.


git-svn-id: https://www.sat.ltu.se/svn/rt/arts/trunk@7466 aaf1aab0-4228-0410-ad68-8dceda47f409