* arts-2-1-116
* CMakeLists.txt:

Handle compiler options more gracefully. Instead of hardcoding them
for different compilers, check if the current compiler supports the
flag. If not, just ignore it.

Set and check flags separately for C and C++ compiler.

For Xcode project, pass active configuration name to ctest so it can
select the correct executable.

* cmake/modules/ArtsAddCompilerFlag.cmake: Added.

* src/libmicrohttpd/CMakeLists.txt, src/CMakeLists.txt:

Little tweak for clang.


git-svn-id: https://www.sat.ltu.se/svn/rt/arts/trunk@7616 aaf1aab0-4228-0410-ad68-8dceda47f409