* arts-2-1-131
* src/methods.cc

Add generic input 'dimension' to append functions. Passing in
'leading' means to append to the leftmost dimension (e.g. rows for
Matrix), 'trailing' to the rightmost (==cols).

Add support to append single elements to arrays.

Add support for Matrix.

* src/groups.cc, src/wsv_aux.h (get_array_groups_as_string):

Extend to return only Arrays whose basetype is a group and to return
names of the basetypes of arrays only.

* src/m_append.h:

Implement append for Matrix.

git-svn-id: https://www.sat.ltu.se/svn/rt/arts/trunk@7665 aaf1aab0-4228-0410-ad68-8dceda47f409