* arts-2-1-166
* ChangeLog: Cleanup.

Remove arts-2-1-162 attributed to me (but not done by me) which was a
duplicate entry of arts-2-1-164.

Restore version number arts-2-1-162 of Patrick's last commit which was
falsely changed to arts-2-1-163.

Restore the commit date for arts-2-1-161 to 2012-06-24.

Add a dummy entry for the now non-existent arts-2-1-163 to get the
version numbering back in line.

git-svn-id: https://www.sat.ltu.se/svn/rt/arts/trunk@7708 aaf1aab0-4228-0410-ad68-8dceda47f409