* arts-2-1-167
* doc/doxygen/CMakeLists.txt:

Rebuild of the documentation was not triggered after source changes.
Add dependencies to all source files. Remove target apidoc_forced.

Remove latex related code because we don't build it anyway.

* doc/doxygen/Doxyfile.in: Update config file to doxygen 1.7.

Add MARKDOWN_SUPPORT=NO. On doxygen 1.8, the tags on
the mainpage are not substituted properly. Currently commented out
because it generates a warning on earlier doxygen versions.

git-svn-id: https://www.sat.ltu.se/svn/rt/arts/trunk@7710 aaf1aab0-4228-0410-ad68-8dceda47f409