* arts-2-1-288
* src/montecarlo.cc (mcPathTraceIPA):

Add an assertion on istep to make sure we ran at least once through
the while loop. If not, several variables are used uninitialized.

* src/cloudbox.cc (scale_H11):

Move definition of ratio into the block where it's actually used and
remove unnecessary assignment.

(pnd_fieldMP48): Init lambda with nan and add an assertion to make sure
it was properly initialized inside the while loop. Otherwise it's
possible that it is used uninitialized.

* src/ppath.cc (ppath_start_2d, do_gridcell_2d):

Remove unnecessary assignments to variables whose value is never read
again.

* src/m_atmosphere.cc (z_fieldFromHSE):

Remove unnecessary assignment of g1=g0 by moving it inside the
for-loop where it is actually used (and initialized with g2).

git-svn-id: https://www.sat.ltu.se/svn/rt/arts/trunk@7860 aaf1aab0-4228-0410-ad68-8dceda47f409