* arts-2-1-294
* src/absorption.{cc,h} (checkIsotopologueRatios):

Add function to validate isotopologue ratios.

* src/m_abs.cc (abs_mat_per_speciesAddOnTheFly), src/m_abs_lookup.cc
(abs_lookupCreate), src/m_zeeman.cc (abs_mat_per_speciesAddZeeman):

Call checkIsotopologueRatios. Probably not the best places to call it,
because it is called more often than needed. Other option is to make a
workspace method that has to be called by the user. At least this way,
the check is automatic. Something to be discussed.

git-svn-id: https://www.sat.ltu.se/svn/rt/arts/trunk@7869 aaf1aab0-4228-0410-ad68-8dceda47f409