* arts-2-1-299
* This commit fixes the behaviour of the TYPECreate methods. After
creating a new variable, instead of being initialized with 0 or empty,
its state is now uninitialized. This is the same as for the builtin
workspace variables.

* src/agenda_class.h (MRecord):

Add minternal flag. Is set to false if the method is from the
controlfile and true if it was added by the engine. For example, Create
and Delete calls for automatically generated variables (for literals)
are flagged as internal.

* src/agenda_class.cc (Agenda::execute):

Use the new isInternal flag to determine output level instead of some
funky string comparisons on the method and variables names.

* src/methods.cc (define_md_data_raw):

Update docs for the Create methods.

* src/parser.cc (ArtsParser::parse_agenda):

Insert a Delete after TYPECreate methods to make the new variable
appear uninitialized.

(ArtsParser::tasklist_insert_set_delete):

Set the internal flag for MRecord to true for the automatically
inserted Set and Delete methods.

git-svn-id: https://www.sat.ltu.se/svn/rt/arts/trunk@7874 aaf1aab0-4228-0410-ad68-8dceda47f409